Move mobile feature adapter onto SDK contracts
