+++ src/Honua.Mobile.Maui/HonuaMobileServiceCollectionExtensions.cs
@@ -11,6 +11,7 @@ using Honua.Mobile.Sdk.Scenes;
 using Honua.Sdk.Abstractions.Features;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
+using SdkFeatureClient = Honua.Mobile.Sdk.Features.HonuaMobileSdkFeatureClient;
 using SdkOfflineChangeJournal = Honua.Sdk.Offline.Abstractions.IOfflineChangeJournal;
 using SdkOfflineFeatureStore = Honua.Sdk.Offline.Abstractions.IOfflineFeatureStore;
 using SdkOfflinePackageManifest = Honua.Sdk.Offline.Abstractions.OfflinePackageManifest;
@@ -169,9 +170,9 @@ public static class HonuaMobileServiceCollectionExtensions
         services.AddSingleton<IGeoPackageSyncStore, GeoPackageSyncStore>();
         services.AddSingleton<IConnectivityStateProvider, AlwaysOnlineConnectivityStateProvider>();
         services.AddSingleton<GeoPackageSdkOfflineStoreAdapter>();
-        services.AddSingleton<HonuaMobileSdkFeatureClient>();
-        services.AddSingleton<IHonuaFeatureQueryClient>(sp => sp.GetRequiredService<HonuaMobileSdkFeatureClient>());
-        services.AddSingleton<IHonuaFeatureEditClient>(sp => sp.GetRequiredService<HonuaMobileSdkFeatureClient>());
+        services.AddSingleton<SdkFeatureClient>();
+        services.AddSingleton<IHonuaFeatureQueryClient>(sp => sp.GetRequiredService<SdkFeatureClient>());
+        services.AddSingleton<IHonuaFeatureEditClient>(sp => sp.GetRequiredService<SdkFeatureClient>());
         services.AddSingleton<SdkOfflineFeatureStore>(sp => sp.GetRequiredService<GeoPackageSdkOfflineStoreAdapter>());
         services.AddSingleton<SdkOfflineChangeJournal>(sp => sp.GetRequiredService<GeoPackageSdkOfflineStoreAdapter>());
         services.AddSingleton<SdkOfflineCheckpointStore>(sp => sp.GetRequiredService<GeoPackageSdkOfflineStoreAdapter>());
