+++ src/Honua.Mobile.Offline/Honua.Mobile.Offline.csproj
@@ -6,8 +6,9 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Honua.Sdk.Offline" Version="0.1.0-alpha.1" />
-    <PackageReference Include="Honua.Sdk.Offline.Abstractions" Version="0.1.0-alpha.1" />
+    <PackageReference Include="Honua.Sdk.Abstractions" Version="0.1.2-alpha.1" />
+    <PackageReference Include="Honua.Sdk.Offline" Version="0.1.2-alpha.1" />
+    <PackageReference Include="Honua.Sdk.Offline.Abstractions" Version="0.1.2-alpha.1" />
     <PackageReference Include="Microsoft.Data.Sqlite" Version="10.0.0" />
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="10.0.0" />
   </ItemGroup>
