+++ src/Honua.Mobile.Sdk/Honua.Mobile.Sdk.csproj
@@ -18,6 +18,7 @@
 
   <ItemGroup>
     <None Include="..\..\README.md" Pack="true" PackagePath="\" />
+    <PackageReference Include="Honua.Sdk.Abstractions" Version="0.1.2-alpha.1" />
     <PackageReference Include="Google.Protobuf" Version="3.33.1" />
     <PackageReference Include="Grpc.Core.Api" Version="2.71.0" />
     <PackageReference Include="Grpc.Net.Client" Version="2.71.0" />
